project--azure-document-cataloging/requirements.txt
openai
azure-ai-formrecognizer
python-dotenv
tiktoken